pyproject.toml
[project]
name = "geopolitical-risk-ingestion-pipeline"
version = "0.9.0"
description = "DENG ingestion pipeline with bronze, silver, and gold analytical layers"
readme = "README.md"
license = "AGPL-3.0-or-later"
license-files = ["LICENSE"]
classifiers = [
  "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
]
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "psycopg2-binary>=2.9.11",
    "python-dateutil>=2.9.0",
    "python-dotenv>=1.2.2",
]
[dependency-groups]
dev = [
    "black>=26.1.0",
    "jupyter>=1.1.1",
    "mypy>=1.20.0",
    "pandas>=3.0.1",
    "pgcli>=4.4.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.15.9",
]

[tool.uv]
default-groups = []

[project.scripts]
deng-ingestion = "deng_ingestion.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/deng_ingestion"]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.venv
  | \.git
)/
'''

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "B",
  "UP",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true

[tool.pytest.ini_options]
pythonpath = ["src"]
